pyproject.toml
[project]
name = "datatap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.0",
    "flask>=3.1.3",
    "httpx>=0.28.1",
    "pydantic>=2.13.2",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]
